103-be/shibboleth-login-wrong-error-page (#279)
* The user is redireced to the FE after wrong shibboleth authentication.

* Changed redirect URLs

* Fixed checkstyle issues

* Copied ShibbolethFilterIT into ClarinShibbolethFilterIT because the tests did must be updated.
